Tests/LibWeb/Text/expected/UIEvents/inputevent-gettargetranges.txt

Target range count: 1. Start offset: 1, End offset: 3
No target ranges
